+++ aiida_kkr/workflows/voro_start.py
@@ -11,7 +11,7 @@ from aiida.work.launch import submit
 from aiida.work import workfunction as wf
 from aiida_kkr.calculations.kkr import KkrCalculation
 from aiida_kkr.calculations.voro import VoronoiCalculation
-from aiida_kkr.tools.kkr_params import kkrparams
+from masci_tools.io.kkr_params import kkrparams
 from aiida_kkr.workflows.dos import kkr_dos_wc
 from aiida_kkr.tools.common_workfunctions import (test_and_get_codenode, update_params, 
                                                   update_params_wf, get_inputs_voronoi)